requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
pysnmp==4.4.12
pyasn1==0.4.8
